add a color picker and add MainDonut component
