+++ package.json
@@ -26,6 +26,7 @@
     "framer-motion": "^10.16.4",
     "next": "13.5.2",
     "react": "18.2.0",
+    "react-colorful": "^5.6.1",
     "react-dom": "18.2.0",
     "react-slick": "^0.29.0",
     "reactjs-popup": "^2.0.6",
